Move components from Register-Login Page to RegisterLoginPage directory
