+++ src/Headers&Footers/Header.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import '../styles/Header.css';
 import { HeaderState } from '../types/Header.d';
-import Login from '../Register-Login Page/Login';
-import Register from '../Register-Login Page/Register';
+import Login from '../RegisterLoginPage/Login';
+import Register from '../RegisterLoginPage/Register';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { showRegisterWindow, showLoginWindow } from '../actions/appActions';
